Move widgets from GUIslice_ex into individual files #120
